create,delete aun update-saver rules whit node ready

+++ api/routes/devices.js
@@ -23,9 +23,8 @@ const auth = {
     password: "public"
   }
 };
-//===============GET-DEVICE========================
+///GET DEVICES
 router.get("/device", checkAuth, async (req, res) => {
-
   try {
     const userId = req.userData._id;
     const devices = await Device.find({ userId: userId });
@@ -34,11 +33,10 @@ router.get("/device", checkAuth, async (req, res) => {
       status: "success",
       data: devices
     };
+
     res.json(toSend);
-
   } catch (error) {
-
-    console.log("ERROR GETTING DEVICES")
+    console.log("ERROR GETTING DEVICES");
 
     const toSend = {
       status: "error",
@@ -48,7 +46,8 @@ router.get("/device", checkAuth, async (req, res) => {
     return res.status(500).json(toSend);
   }
 });
-//===============NEW-DEVICE========================
+
+//NEW DEVICE
 router.post("/device", checkAuth, async (req, res) => {
   try {
     const userId = req.userData._id;
@@ -58,6 +57,8 @@ router.post("/device", checkAuth, async (req, res) => {
     newDevice.createdTime = Date.now();
 
     const device = await Device.create(newDevice);
+
+    createSaverRule(userId,newDevice.dId, true);
 
     selectDevice(userId, newDevice.dId);
 
@@ -78,25 +79,22 @@ router.post("/device", checkAuth, async (req, res) => {
     return res.status(500).json(toSend);
   }
 });
-//===============DELETE-DEVICE=====================
-router.delete("/device", checkAuth, async(req, res) => {
 
+//DELETE DEVICE
+router.delete("/device", checkAuth, async (req, res) => {
   try {
     const userId = req.userData._id;
     const dId = req.query.dId;
 
-    const result = await Device.deleteOne({userId:
-       userId, dId: dId});
-  
+    const result = await Device.deleteOne({ userId: userId, dId: dId });
+
     const toSend = {
       status: "success",
       data: result
     };
-  
-    return res.json(toSend);
-    
-  } catch (error) {
 
+    return res.json(toSend);
+  } catch (error) {
     console.log("ERROR DELETING DEVICE");
     console.log(error);
 
@@ -108,7 +106,8 @@ router.delete("/device", checkAuth, async(req, res) => {
     return res.status(500).json(toSend);
   }
 });
-//================UPDATE DEVICE====================
+
+//UPDATE DEVICE
 router.put("/device", checkAuth, (req, res) => {
   const dId = req.body.dId;
   const userId = req.userData._id;
@@ -137,9 +136,7 @@ ______ _   _ _   _ _____ _____ _____ _____ _   _  _____
 \_|    \___/\_| \_/\____/ \_/  \___/ \___/\_| \_/\____/  
 */
 
-setTimeout(() => {
-  createSaverRule("121212","11111",false);
-}, 2000);
+
 
 async function selectDevice(userId, dId) {
   try {
@@ -160,68 +157,110 @@ async function selectDevice(userId, dId) {
     return false;
   }
 }
+
 /*
  SAVER RULES FUNCTIONS
 */
-//get saver rule
+//get saver rules
+async function getSaverRules(userId) {
+  try {
+    const rules = await SaverRule.find({ userId: userId });
+    return rules;
+  } catch (error) {
+    return false;
+  }
+}
 
 //create saver rule
 async function createSaverRule(userId, dId, status) {
-
   try {
     const url = "http://localhost:8085/api/v4/rules";
 
-  const topic = userId + "/" + dId + "/+/sdata";
+    const topic = userId + "/" + dId + "/+/sdata";
 
-  const rawsql = "SELECT topic, payload FROM \"" + topic + "\" WHERE payload.save = 1";
+    const rawsql =
+      'SELECT topic, payload FROM "' + topic + '" WHERE payload.save = 1';
 
-  var newRule = {
-    rawsql: rawsql,
-    actions: [
-      {
-        name: "data_to_webserver",
-        params: {
-          $resource: global.saverResource.id,
-          payload_tmpl: '{"userId":"' +  userId + '","payload":${payload},"topic":"${topic}"}'
+    var newRule = {
+      rawsql: rawsql,
+      actions: [
+        {
+          name: "data_to_webserver",
+          params: {
+            $resource: global.saverResource.id,
+            payload_tmpl:
+              '{"userId":"' +
+              userId +
+              '","payload":${payload},"topic":"${topic}"}'
+          }
         }
-      }
-    ],
-    description: "SAVER-RULE",
-    enabled: status
-  };
+      ],
+      description: "SAVER-RULE",
+      enabled: status
+    };
 
-  //save rule in emqx - grabamos la regla en emqx
-  const res = await axios.post(url, newRule, auth);
+    //save rule in emqx - grabamos la regla en emqx
+    const res = await axios.post(url, newRule, auth);
 
-  if(res.status === 200 && res.data.data){
-    console.log(res.data.data);
+    if (res.status === 200 && res.data.data) {
+      console.log(res.data.data);
 
-    await SaverRule.create({
-      userId: userId,
-      dId: dId,
-      emqxRuleId: res.data.data.id,
-      status: status
-    });
+      await SaverRule.create({
+        userId: userId,
+        dId: dId,
+        emqxRuleId: res.data.data.id,
+        status: status
+      });
 
-    return true;
-
-  }else{
-    return false;
-  }
-
+      return true;
+    } else {
+      return false;
+    }
   } catch (error) {
-    console.log("Error creating saver rule")
+    console.log("Error creating saver rule");
     console.log(error);
     return false;
   }
-
-
 }
 
 //update saver rule
+async function updateSaverRuleStatus(emqxRuleId, status) {
+  const url = "http://localhost:8085/api/v4/rules/" + emqxRuleId;
+
+  const newRule = {
+    enabled: status
+  };
+
+  const res = await axios.put(url, newRule, auth);
+
+  if (res.status === 200 && res.data.data) {
+    await SaverRule.updateOne({ emqxRuleId: emqxRuleId }, { status: status });
+    console.log("Saver Rule Status Updated...".green);
+    return {
+      status: "success",
+      action: "updated"
+    };
+  }
+}
 
 //delete saver rule
+async function deleteSaverRule(dId) {
+  try {
+    const mongoRule = await SaverRule.findOne({ dId: dId });
 
+    const url = "http://localhost:8085/api/v4/rules/" + mongoRule.emqxRuleId;
+
+    const emqxRule = await axios.delete(url, auth);
+
+    const deleted = await SaverRule.deleteOne({ dId: dId });
+
+    return true;
+  } catch (error) {
+    console.log("Error deleting saver rule");
+    console.log(error);
+    return false;
+  }
+}
 
 module.exports = router;
